chore: consolidate packages/design-lint into packages/eslint-config (per user direction)
Single source of truth for design-system ESLint rules. Per user direction
(2026-04-29), packages/design-lint is folded into the existing
packages/eslint-config so we don't ship two competing lint packages.

Reconciled rule differences (kept superset):
- no-hardcoded-colors: adopted design-lint's broader version (catches all
  Literal/TemplateElement/JSX, not just JSX style/className).
- no-raw-spacing: adopted design-lint's broader version (full spacing prop
  set, var() detection, picks up `const styles = {}` not just JSX style).
- prefer-token-import: merged both — kept eslint-config's legacy
  @oneimpression/design-tokens patterns AND added design-lint's product-
  local color/palette/theme suspect detection.
- Kept eslint-config's other 4 rules unchanged: no-inline-styles,
  no-raw-surface, no-hardcoded-radius, no-hardcoded-typography.

Additions to eslint-config:
- strict.js preset (every rule = error) ported from design-lint.
- ./recommended and ./strict subpath exports.
- node:test test suite covering all 7 rules (extends design-lint's 3-rule
  suite). All 7 pass.
- eslint devDep + test script.
- Version bumped 2.0.1 -> 2.1.0.

Other:
- Deleted packages/design-lint/ entirely.
- Updated CI workflow step "Test design-lint rules" -> "Test eslint-config
  rules" pointing to new test path.
- Regenerated package-lock.json (no design-lint workspace entry).

Tests: all 7 eslint-config rule tests pass via `npm test -w
packages/eslint-config`. Workspace build clean. Pre-existing mcp-server
extract-props test failure is unrelated to this change.
