--- packages/design-lint/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@amplify/design-lint",
-  "version": "1.0.0",
-  "description": "Amplify Design System — ESLint flat-config that enforces token usage (no-hardcoded-colors, no-raw-spacing, prefer-token-import)",
-  "main": "index.js",
-  "exports": {
-    ".": "./index.js",
-    "./recommended": "./index.js",
-    "./strict": "./strict.js"
-  },
-  "scripts": {
-    "test": "node --test test/*.test.js"
-  },
-  "peerDependencies": {
-    "eslint": ">=9.0.0"
-  },
-  "devDependencies": {
-    "eslint": "^9.0.0"
-  },
-  "files": [
-    "index.js",
-    "strict.js",
-    "rules/"
-  ]
-}
