+++ packages/eslint-config/index.js
@@ -1,5 +1,5 @@
 /**
- * @amplify-ai/eslint-config — Shared ESLint flat config
+ * @amplify-ai/eslint-config — Shared ESLint flat config (recommended preset)
  *
  * Enforces Amplify design token usage across consuming projects.
  *
@@ -10,7 +10,12 @@
  * Or CommonJS:
  *   const amplifyConfig = require('@amplify-ai/eslint-config');
  *   module.exports = [...amplifyConfig];
+ *
+ * For hard CI enforcement (every violation = error), use the strict preset:
+ *   import amplifyStrict from '@amplify-ai/eslint-config/strict';
+ *   export default [...amplifyStrict];
  */
+'use strict';
 
 const noHardcodedColors = require('./rules/no-hardcoded-colors');
 const noRawSpacing = require('./rules/no-raw-spacing');
@@ -23,7 +28,7 @@ const noHardcodedTypography = require('./rules/no-hardcoded-typography');
 const plugin = {
   meta: {
     name: '@amplify-ai/eslint-plugin',
-    version: '2.0.0',
+    version: '2.1.0',
   },
   rules: {
     'no-hardcoded-colors': noHardcodedColors,
@@ -38,6 +43,7 @@ const plugin = {
 
 module.exports = [
   {
+    files: ['**/*.{js,jsx,ts,tsx,mjs,cjs}'],
     plugins: {
       '@amplify-ai': plugin,
     },
@@ -52,3 +58,5 @@ module.exports = [
     },
   },
 ];
+
+module.exports.plugin = plugin;
