+++ packages/eslint-config/rules/no-hardcoded-colors.js
@@ -1,93 +1,63 @@
 /**
- * Rule: no-hardcoded-colors
+ * @amplify-ai/eslint-config — no-hardcoded-colors
  *
- * Warns when inline hex color values are used in JSX style props or
- * className strings. Encourages use of design token CSS variables instead.
+ * Flags hex colour literals (#fff, #ff0000, #FF000080) in:
+ *   - JSX style props: style={{ color: '#FF0000' }}
+ *   - className strings: className="text-[#abc]"
+ *   - JS object properties: const styles = { color: '#abc' }
+ *   - Tagged template literals (CSS-in-JS): css`color: #abc;`
+ *   - Plain CSS string literals
  *
- * Catches patterns like:
- *   - style={{ color: '#FF0000' }}
- *   - style={{ backgroundColor: '#abc' }}
- *   - className="text-[#FF0000]"
+ * Acceptable: CSS variables (var(--amplify-*) / var(--amp-*)) or token JS imports.
  */
+'use strict';
 
-const HEX_COLOR_REGEX = /#(?:[0-9a-fA-F]{3,4}){1,2}\b/;
+// Match: #abc, #abcd, #aabbcc, #aabbccdd
+const HEX_RE = /#(?:[0-9a-fA-F]{3,4}\b|[0-9a-fA-F]{6}\b|[0-9a-fA-F]{8}\b)/;
+
+function checkString(context, node, value) {
+  if (typeof value !== 'string') return;
+  const m = value.match(HEX_RE);
+  if (!m) return;
+  context.report({
+    node,
+    messageId: 'noHardcodedColor',
+    data: { color: m[0] },
+  });
+}
 
 module.exports = {
   meta: {
     type: 'suggestion',
     docs: {
-      description: 'Disallow hardcoded hex color values in JSX; use design tokens instead',
+      description:
+        'Disallow hardcoded hex colour literals; require Amplify design tokens (CSS vars or @amplify-ai/tokens-* imports)',
       recommended: true,
     },
+    schema: [],
     messages: {
       noHardcodedColor:
-        'Avoid hardcoded color "{{color}}". Use an Amplify design token (CSS variable or JS import) instead.',
+        'Hardcoded color "{{color}}" — use an Amplify design token (--amplify-* / --amp-* CSS variable) instead.',
     },
-    schema: [],
   },
-
   create(context) {
     return {
-      // Catches style={{ color: '#FFF' }} in JSX
-      JSXAttribute(node) {
-        if (node.name.name !== 'style') return;
-
-        const expr = node.value && node.value.expression;
-        if (!expr || expr.type !== 'ObjectExpression') return;
-
-        for (const prop of expr.properties) {
-          if (
-            prop.value &&
-            prop.value.type === 'Literal' &&
-            typeof prop.value.value === 'string'
-          ) {
-            const match = prop.value.value.match(HEX_COLOR_REGEX);
-            if (match) {
-              context.report({
-                node: prop.value,
-                messageId: 'noHardcodedColor',
-                data: { color: match[0] },
-              });
-            }
-          }
+      Literal(node) {
+        if (typeof node.value !== 'string') return;
+        // Skip module imports / dependency paths
+        const parent = node.parent;
+        if (parent && (parent.type === 'ImportDeclaration' || parent.type === 'ExportAllDeclaration')) return;
+        checkString(context, node, node.value);
+      },
+      TemplateElement(node) {
+        if (node.value && node.value.cooked) {
+          checkString(context, node, node.value.cooked);
         }
       },
-
-      // Catches className="text-[#FF0000]" or template literals with hex colors
       JSXAttribute(node) {
-        if (node.name.name !== 'className') return;
-
-        if (node.value && node.value.type === 'Literal' && typeof node.value.value === 'string') {
-          const match = node.value.value.match(HEX_COLOR_REGEX);
-          if (match) {
-            context.report({
-              node: node.value,
-              messageId: 'noHardcodedColor',
-              data: { color: match[0] },
-            });
-          }
-        }
-      },
-
-      // Catches hex colors in template literals used in JSX attributes
-      TemplateLiteral(node) {
-        for (const quasi of node.quasis) {
-          const match = quasi.value.raw.match(HEX_COLOR_REGEX);
-          if (match) {
-            // Only report if this template literal is inside a JSX attribute
-            let parent = node.parent;
-            while (parent) {
-              if (parent.type === 'JSXAttribute') {
-                context.report({
-                  node: quasi,
-                  messageId: 'noHardcodedColor',
-                  data: { color: match[0] },
-                });
-                break;
-              }
-              parent = parent.parent;
-            }
-          }
+        // className="text-[#abc]" via JSX literal
+        if (node.value && node.value.type === 'Literal') {
+          checkString(context, node.value, node.value.value);
         }
       },
     };
