+++ packages/eslint-config/rules/no-raw-spacing.js
@@ -1,31 +1,95 @@
 /**
  * @amplify-ai/eslint-config — no-raw-spacing
- * Warns when raw pixel values are used in style props where spacing tokens exist.
+ *
+ * Flags raw pixel/em/rem literals used as spacing in JSX style props or
+ * style objects, e.g.:
+ *   style={{ padding: '12px' }}            // BAD
+ *   style={{ margin: 16 }}                  // BAD
+ *   const x = { paddingTop: '0.5rem' }      // BAD
+ *
+ * Allowed:
+ *   style={{ padding: 'var(--amp-spacing-md)' }}
+ *   style={{ padding: tokens.spacing.md }}
+ *   numeric 0 (semantic for "no spacing")
  */
+'use strict';
+
+const SPACING_PROPS = new Set([
+  'padding', 'paddingTop', 'paddingRight', 'paddingBottom', 'paddingLeft',
+  'paddingInline', 'paddingBlock', 'paddingInlineStart', 'paddingInlineEnd',
+  'paddingBlockStart', 'paddingBlockEnd',
+  'margin', 'marginTop', 'marginRight', 'marginBottom', 'marginLeft',
+  'marginInline', 'marginBlock', 'marginInlineStart', 'marginInlineEnd',
+  'marginBlockStart', 'marginBlockEnd',
+  'gap', 'rowGap', 'columnGap',
+  'top', 'right', 'bottom', 'left',
+  'inset', 'insetInline', 'insetBlock',
+]);
+
+const PX_RE = /^-?\d+(?:\.\d+)?(?:px|rem|em)?$/;
+
+function checkPropValue(context, node, propName, value) {
+  if (!SPACING_PROPS.has(propName)) return;
+
+  if (typeof value === 'number') {
+    if (value === 0) return;
+    context.report({ node, messageId: 'noRawSpacing', data: { prop: propName, value: String(value) } });
+    return;
+  }
+
+  if (typeof value !== 'string') return;
+  if (value.startsWith('var(')) return;
+  if (value.includes('--')) return; // Likely embedded CSS var
+  // Skip percent / vw / vh / fr / auto / etc.
+  if (PX_RE.test(value.trim())) {
+    context.report({ node, messageId: 'noRawSpacing', data: { prop: propName, value } });
+  }
+}
+
 module.exports = {
   meta: {
     type: 'suggestion',
-    docs: { description: 'Disallow raw pixel values in style props — use spacing tokens instead' },
-    messages: { noRawSpacing: 'Avoid raw pixel value "{{ value }}". Use a spacing token.' },
+    docs: {
+      description:
+        'Disallow raw px/rem/em spacing literals in style objects; use Amplify spacing tokens instead',
+      recommended: true,
+    },
     schema: [],
+    messages: {
+      noRawSpacing:
+        'Raw spacing value `{{value}}` for `{{prop}}` — use a token like var(--amp-spacing-md) or @amplify-ai/tokens-foundation/js.',
+    },
   },
   create(context) {
-    const SPACING_MAP = {
-      '4px': 'spacing.1', '8px': 'spacing.2', '12px': 'spacing.3',
-      '16px': 'spacing.4', '20px': 'spacing.5', '24px': 'spacing.6',
-      '32px': 'spacing.8', '40px': 'spacing.10', '48px': 'spacing.12',
-      '64px': 'spacing.16', '80px': 'spacing.20', '96px': 'spacing.24',
-    };
+    function visitObjectExpression(obj) {
+      if (!obj || obj.type !== 'ObjectExpression') return;
+      for (const prop of obj.properties) {
+        if (prop.type !== 'Property') continue;
+        const keyName =
+          prop.key.type === 'Identifier' ? prop.key.name :
+          prop.key.type === 'Literal' ? String(prop.key.value) : null;
+        if (!keyName) continue;
+        const v = prop.value;
+        if (v.type === 'Literal') {
+          checkPropValue(context, v, keyName, v.value);
+        } else if (v.type === 'TemplateLiteral' && v.expressions.length === 0) {
+          checkPropValue(context, v, keyName, v.quasis.map((q) => q.value.cooked).join(''));
+        }
+      }
+    }
+
     return {
+      // style={{ ... }}
       JSXAttribute(node) {
-        if (node.name.name !== 'style') return;
-        const value = node.value;
-        if (!value || value.type !== 'JSXExpressionContainer') return;
-        const source = context.getSourceCode().getText(value);
-        for (const [px, token] of Object.entries(SPACING_MAP)) {
-          if (source.includes(`'${px}'`) || source.includes(`"${px}"`)) {
-            context.report({ node, messageId: 'noRawSpacing', data: { value: `${px} → use ${token}` } });
-          }
+        if (node.name && node.name.name === 'style') {
+          const expr = node.value && node.value.expression;
+          visitObjectExpression(expr);
+        }
+      },
+      // const styles = { padding: '12px' }
+      VariableDeclarator(node) {
+        if (node.init && node.init.type === 'ObjectExpression') {
+          visitObjectExpression(node.init);
         }
       },
     };
